breakdown accountSettings duck
