+++ src/features/Account/GlobalSettings.tsx
@@ -9,7 +9,7 @@ import CircleProgress from 'src/components/CircleProgress';
 import { StyleRulesCallback, Theme, withStyles, WithStyles } from 'src/components/core/styles';
 import ErrorState from 'src/components/ErrorState';
 import TagImportDrawer from 'src/features/TagImport';
-import { updateAccountSettings } from 'src/store/accountSettings';
+import { updateAccountSettings } from 'src/store/accountSettings/accountSettings.requests';
 import { handleOpen } from 'src/store/backupDrawer';
 import getEntitiesWithGroupsToImport, { emptyGroupedEntities, GroupedEntitiesForImport } from 'src/store/selectors/getEntitiesWithGroupsToImport';
 import { openGroupDrawer } from 'src/store/tagImportDrawer';
