+++ src/store/backupDrawer/index.ts
@@ -3,7 +3,7 @@ import { isEmpty, pathOr } from 'ramda';
 import { Reducer } from 'redux';
 import { updateAccountSettings } from 'src/services/account';
 import { enableBackups } from 'src/services/linodes';
-import { handleUpdate } from 'src/store/accountSettings';
+import { handleUpdate } from 'src/store/accountSettings/accountSettings.actions';
 import { updateMultipleLinodes } from 'src/store/linodes/linodes.actions';
 import { ThunkActionCreator } from '../types';
